requirements.txt
flask
pandas
numpy
sentence-transformers
faiss-cpu
requests
aiohttp
